requirements.txt
attrs==23.2.0
black==24.2.0
build==1.0.3
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
coverage==7.4.4
coverage-badge==1.1.0
cryptography==42.0.5
distlib==0.3.8
dlint==0.14.1
docutils==0.20.1
dparse==0.6.3
eradicate==2.3.0
exceptiongroup==1.2.0
filelock==3.13.3
flake8==7.0.0
flake8-comprehensions==3.14.0
flake8-eradicate==1.5.0
Flake8-pyproject==1.2.3
flake8-spellcheck==0.28.0
flake8-typing-imports==1.15.0
identify==2.5.35
idna==3.6
importlib_metadata==7.1.0
iniconfig==2.0.0
isort==5.13.2
jaraco.classes==3.4.0
jaraco.context==4.3.0
jaraco.functools==4.0.0
jeepney==0.8.0
keyring==25.1.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.2.0
mypy==1.8.0
mypy-extensions==1.0.0
nh3==0.2.17
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
pep8-naming==0.13.3
pkginfo==1.10.0
platformdirs==4.2.0
pluggy==1.4.0
pre-commit==3.6.0
pycodestyle==2.11.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.17.2
pyproject_hooks==1.0.0
pytest==7.4.4
pytest-mock==3.12.0
pytest-watcher==0.4.2
PyYAML==6.0.1
readme_renderer==43.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safety==2.3.4
SecretStorage==3.3.3
tomli==2.0.1
twine==4.0.2
types-PyYAML==6.0.12.20240311
typing_extensions==4.10.0
urllib3==2.2.1
virtualenv==20.25.1
watchdog==4.0.0
zipp==3.18.1
